bpf: add deny policy enforcement
This change adds the deny policy enforcement in the datapath. The logic
will be the same as done for the "allow" case, i.e. for the policy map
that allows endpoints to send / receive packets. However, the decision
of the logic will be inverted, i.e. if an identity is present in the
deny map, the packets should be immediately dropped by the datapath.

Signed-off-by: André Martins <andre@cilium.io>