pkg/policy: use PortDenyRules for deny policies
Since not all functionalities will be available in the IngressDeny and
EgressDeny fields of Rule, it was necessary to create a new structure,
IngressDeny and EgressDeny. These 2 structures share common fields with
the Ingress and Egress except the implementation of L7 Policies, which
is part of the PortRule but not of the PortDenyRules. This requirement
requires the introduction of some interfaces to avoid code duplication
in the policy calculation.

Signed-off-by: André Martins <andre@cilium.io>